move fetch metrics to separated module
