+++ src/bot/metrics/getDudCount.js
@@ -1,0 +1,3 @@
+const getDudCount = (text) => (/\?/.test(text) ? 1 : 0);
+
+module.exports = getDudCount;
