+++ src/bot/metrics/getPhotoCount.js
@@ -1,0 +1,3 @@
+const getPhotoCount = (photo, text) => ((photo || /.*.jpg|.*.png/.test(text)) ? 1 : 0);
+
+module.exports = getPhotoCount;
