+++ src/bot/metrics/getStickySetName.js
@@ -1,0 +1,3 @@
+const getStickySetName = (sticker) => (sticker && sticker.set_name ? sticker.set_name : '');
+
+module.exports = getStickySetName;
