+++ src/bot/metrics/getTextLength.js
@@ -1,0 +1,3 @@
+const getTextLength = (text) => (text ? text.length : 0);
+
+module.exports = getTextLength;
