+++ src/bot/metrics/getVideoCount.js
@@ -1,0 +1,3 @@
+const getVideoCount = (video, text) => ((video || /.*youtu.be*|.*youtube.com*/.test(text)) ? 1 : 0);
+
+module.exports = getVideoCount;
