+++ src/bot/metrics/getVoiceCount.js
@@ -1,0 +1,3 @@
+const getVoiceCount = (voice) => (voice ? 1 : 0);
+
+module.exports = getVoiceCount;
